"tolto code smells"

+++ src/main/java/org/application/gameshelfapp/buyvideogames/entities/CatalogueDAOCSV.java
@@ -6,7 +6,10 @@ import com.opencsv.exceptions.CsvValidationException;
 import org.application.gameshelfapp.login.exception.PersistencyErrorException;
 
 import java.io.*;
+import java.nio.channels.FileChannel;
 import java.nio.file.Files;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
 
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 
@@ -14,10 +17,12 @@ public class CatalogueDAOCSV implements CatalogueDAO {
 
     private final String filename;
     private final File fd;
+    private final Lock lock;
 
     public CatalogueDAOCSV() {
         this.filename = "src/main/resources/org/application/gameshelfapp/persistency/catalogue.csv";
         this.fd = new File(this.filename);
+        this.lock = new ReentrantLock();
     }
 
     @Override
@@ -27,10 +32,20 @@ public class CatalogueDAOCSV implements CatalogueDAO {
         myRecord[CatalogueAttributes.USERNAME.ordinal()] = username;
         myRecord[CatalogueAttributes.GAMENAME.ordinal()] = gameName;
         myRecord[CatalogueAttributes.COPIES.ordinal()] = game.getOwnerCopies().toString();
-        File tempFile = File.createTempFile("catalogue", "csv");
+        File tempFile = new File("temp_catalogue.csv");
+
+        if(!tempFile.exists()) {
+            boolean created = tempFile.createNewFile();
+            if(!created) throw new PersistencyErrorException("Couldn't save game");
+        }
+
+        this.lock.lock();
 
         try (CSVWriter csvWriter = new CSVWriter(new BufferedWriter(new FileWriter(tempFile)));
-            CSVReader csvReader = new CSVReader(new BufferedReader(new FileReader(this.fd)))){
+             CSVReader csvReader = new CSVReader(new BufferedReader(new FileReader(this.fd)));
+             FileChannel channel = new FileOutputStream(tempFile).getChannel()){
+
+            channel.truncate(0);
             while((myRecord = csvReader.readNext()) != null){
                 if(myRecord[CatalogueAttributes.GAMENAME.ordinal()].equals(gameName)){
                     int newNumberOfCopies = numberOfCopies + Integer.parseInt(myRecord[CatalogueAttributes.COPIES.ordinal()]);
@@ -41,20 +56,26 @@ public class CatalogueDAOCSV implements CatalogueDAO {
 
             csvWriter.flush();
             Files.move(tempFile.toPath(), this.fd.toPath(), REPLACE_EXISTING);
-            tempFile.delete();
         } catch (IOException | CsvValidationException e) {
             throw new PersistencyErrorException("Couldn't save game");
-        }
+        }finally{this.lock.unlock();}
     }
 
     @Override
     public void removeVideogame(String username, Videogame game, int quantity) throws PersistencyErrorException, IOException {
         String[] myRecord;
-        File tempFile = File.createTempFile("catalogue", "csv");;
+        File tempFile = new File("temp_catalogue");
+        if(!tempFile.exists()) {
+            boolean created = tempFile.createNewFile();
+            if(!created) throw new PersistencyErrorException("Couldn't save game");
+        }
 
+        this.lock.lock();
         try (CSVReader csvReader = new CSVReader(new BufferedReader(new FileReader(this.fd)));
-             CSVWriter csvWriter = new CSVWriter(new BufferedWriter(new FileWriter(tempFile)));){
+             CSVWriter csvWriter = new CSVWriter(new BufferedWriter(new FileWriter(tempFile)));
+             FileChannel channel = new FileOutputStream(tempFile).getChannel()){
 
+            channel.truncate(0);
             while ((myRecord = csvReader.readNext()) != null) {
 
                 if (myRecord[CatalogueAttributes.GAMENAME.ordinal()].equals(game.getName())) {
@@ -68,11 +89,12 @@ public class CatalogueDAOCSV implements CatalogueDAO {
             }
             csvWriter.flush();
             Files.move(tempFile.toPath(), this.fd.toPath(), REPLACE_EXISTING);
-            tempFile.delete();
         } catch (IOException e) {
             throw new PersistencyErrorException("Couldn't remove videogame from catalogue");
         } catch (CsvValidationException e) {
             throw new PersistencyErrorException("Couldn't access to catalogue");
+        }finally{
+            this.lock.unlock();
         }
     }
